Add order confirmation mailer

+++ app/jobs/delete_unpaid_single_orders_job.rb
@@ -3,7 +3,10 @@ class DeleteUnpaidSingleOrdersJob < ApplicationJob
 
   def perform(*args)
   		order = args[0]
+      if order.paid_online == false && order.will_pay_on_delivery == false
       	order.destroy
+      else
      	DeleteUnpaidSingleOrdersJob.set(wait: 10.minutes).perform_later(order)
+     end
   end
 end
